requirements.txt
Django==5.0.11
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
requests==2.32.3
python-dotenv==1.0.1
